debug: restore canary versions from matrix
