+++ .github/workflows/ci_compatibility-deno.yml
@@ -14,11 +14,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        deno-version: [
-            'v1.x',
-            'v2.x',
-            # 'canary'
-          ]
+        deno-version: ['v1.x', 'v2.x', 'canary']
     name: Deno ${{ matrix.deno-version }}
     steps:
       - name: ➕ Actions - Checkout
