+++ test/e2e/before-and-after-each.test.ts
@@ -1,230 +1,43 @@
-import { test } from '../../src/modules/helpers/test.js';
 import { describe } from '../../src/modules/helpers/describe.js';
 import { it } from '../../src/modules/helpers/it/core.js';
-import { poku } from '../../src/modules/essentials/poku.js';
 import { assert } from '../../src/modules/essentials/assert.js';
-import { skip } from '../../src/modules/helpers/skip.js';
-import { GLOBAL } from '../../src/configs/poku.js';
+import { inspectPoku, ext } from '../__utils__/capture-cli.test.js';
 
-if (GLOBAL.runtime === 'deno') skip();
-
-test(async () => {
-  const prepareService = () => new Promise((resolve) => resolve(undefined));
-  const resetService = () => new Promise((resolve) => resolve(undefined));
-  const crashIt = () => new Promise((_, reject) => reject("Let's crash it"));
-  const crashItAgain = () => {
-    throw new Error("Let's crash it");
-  };
-
-  await describe('Before and After Each: direct methods', async () => {
-    await it(async () => {
-      const code = await poku('./test/__fixtures__/e2e/success', {
-        noExit: true,
-        beforeEach: prepareService,
-        afterEach: resetService,
-      });
-
-      assert.strictEqual(
-        code,
-        0,
-        'beforeEach and afterEach hooks with successful path'
-      );
+describe(async () => {
+  await it('Before and After Each File', async () => {
+    const output = await inspectPoku(`-d api.test.${ext}`, {
+      cwd: 'test/__fixtures__/e2e/each-file',
     });
 
-    await it(async () => {
-      const code = await poku('./test/__fixtures__/e2e/fail', {
-        noExit: true,
-        beforeEach: prepareService,
-        afterEach: resetService,
-      });
+    if (output.exitCode !== 0) {
+      console.log(output.stdout);
+      console.log(output.stderr);
+    }
 
-      assert.strictEqual(
-        code,
-        1,
-        'beforeEach and afterEach hooks with failing path'
-      );
-    });
-  });
+    assert.strictEqual(output.exitCode, 0, 'Should Pass');
 
-  await describe('Before and After Each: called methods', async () => {
-    await it(async () => {
-      const code = await poku('./test/__fixtures__/e2e/success', {
-        noExit: true,
-        beforeEach: () => prepareService(),
-        afterEach: () => resetService(),
-      });
+    assert.match(
+      output.stdout,
+      /● Before and After Each: direct methods/,
+      'Before and After Each: direct methods'
+    );
 
-      assert.strictEqual(
-        code,
-        0,
-        'beforeEach and afterEach hooks with successful path'
-      );
-    });
+    assert.match(
+      output.stdout,
+      /● Before and After Each: called methods/,
+      'Before and After Each: called methods'
+    );
 
-    await it(async () => {
-      const code = await poku('./test/__fixtures__/e2e/fail', {
-        noExit: true,
-        beforeEach: () => prepareService(),
-        afterEach: () => resetService(),
-      });
+    assert.match(
+      output.stdout,
+      /● Before and After Each: await called methods/,
+      'Before and After Each: await called methods'
+    );
 
-      assert.strictEqual(
-        code,
-        1,
-        'beforeEach and afterEach hooks with failing path'
-      );
-    });
-  });
-
-  await describe('Before and After Each: await called methods', async () => {
-    await it(async () => {
-      const code = await poku('./test/__fixtures__/e2e/success', {
-        noExit: true,
-        beforeEach: async () => await prepareService(),
-        afterEach: async () => await resetService(),
-      });
-
-      assert.strictEqual(
-        code,
-        0,
-        'beforeEach and afterEach hooks with successful path'
-      );
-    });
-
-    await it(async () => {
-      const code = await poku('./test/__fixtures__/e2e/fail', {
-        noExit: true,
-        beforeEach: async () => await prepareService(),
-        afterEach: async () => await resetService(),
-      });
-
-      assert.strictEqual(
-        code,
-        1,
-        'beforeEach and afterEach hooks with failing path'
-      );
-    });
-  });
-
-  await describe('Before and After Each: anonymous methods', async () => {
-    await it(async () => {
-      const code = await poku('./test/__fixtures__/e2e/success', {
-        noExit: true,
-        beforeEach: () => true,
-        afterEach: () => true,
-      });
-
-      assert.strictEqual(
-        code,
-        0,
-        'beforeEach and afterEach hooks with successful path'
-      );
-    });
-
-    await it(async () => {
-      const code = await poku('./test/__fixtures__/e2e/fail', {
-        noExit: true,
-        beforeEach: () => true,
-        afterEach: () => true,
-      });
-
-      assert.strictEqual(
-        code,
-        1,
-        'beforeEach and afterEach hooks with failing path'
-      );
-    });
-  });
-
-  await describe('Before and After Each: anonymous methods (function)', async () => {
-    await it(async () => {
-      const code = await poku('./test/__fixtures__/e2e/success', {
-        noExit: true,
-        beforeEach: function () {
-          return;
-        },
-        afterEach: function () {
-          return;
-        },
-      });
-
-      assert.strictEqual(
-        code,
-        0,
-        'beforeEach and afterEach hooks with successful path'
-      );
-    });
-
-    await it(async () => {
-      const code = await poku('./test/__fixtures__/e2e/fail', {
-        noExit: true,
-        beforeEach: function () {
-          return;
-        },
-        afterEach: function () {
-          return;
-        },
-      });
-
-      assert.strictEqual(
-        code,
-        1,
-        'beforeEach and afterEach hooks with failing path'
-      );
-    });
-  });
-
-  await describe('Before and After Each: Failure', async () => {
-    await it(async () => {
-      const code = await poku('./test/__fixtures__/e2e/success', {
-        noExit: true,
-        beforeEach: crashIt,
-        afterEach: resetService,
-      });
-
-      assert.strictEqual(
-        code,
-        1,
-        'Rejects beforeEach hook with successful path'
-      );
-    });
-
-    await it(async () => {
-      const code = await poku('./test/__fixtures__/e2e/success', {
-        noExit: true,
-        beforeEach: prepareService,
-        afterEach: crashIt,
-      });
-
-      assert.strictEqual(
-        code,
-        1,
-        'Rejects afterEach hook with successful path'
-      );
-    });
-
-    await it(async () => {
-      const code = await poku('./test/__fixtures__/e2e/success', {
-        noExit: true,
-        beforeEach: crashItAgain,
-        afterEach: resetService,
-      });
-
-      assert.strictEqual(
-        code,
-        1,
-        'Throws beforeEach hook with successful path'
-      );
-    });
-
-    await it(async () => {
-      const code = await poku('./test/__fixtures__/e2e/success', {
-        noExit: true,
-        beforeEach: prepareService,
-        afterEach: crashItAgain,
-      });
-
-      assert.strictEqual(code, 1, 'Throws afterEach hook with successful path');
-    });
+    assert.match(
+      output.stdout,
+      /● Before and After Each: Failure/,
+      'Before and After Each: Failure'
+    );
   });
 });
